move hook in samples provider component to service
