+++ src/services/samples/index.ts
@@ -1,6 +1,7 @@
 import { reducer } from './reducer';
 import * as types from './types';
 import * as actions from './actions';
+import * as hooks from './hooks';
 import fetchSamples from './fetch';
 
-export { actions, fetchSamples, reducer, types };
+export { actions, fetchSamples, hooks, reducer, types };
